LFS-105: UI to browse NCI Thesaurus terms as a tree Move BrowseListChild to its own file, more styling
