+++ modules/thesaurus/src/main/frontend/src/query/browse.jsx
@@ -23,168 +23,9 @@ import { withStyles } from "@material-ui/core/styles";
 import { Dialog, Typography } from '@material-ui/core';
 // material-dashboard-react
 import Button from "material-dashboard-react/dist/components/CustomButtons/Button.js";
-// @material-ui/icons
-import Info from "@material-ui/icons/Info";
 
+import BrowseListChild from "./browseListChild.jsx";
 import BrowseTheme from "./browseStyle.jsx";
-import browseStyle from "./browseStyle.jsx";
-
-class ListChild extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      loadedChildren: false,
-      checkedForChildren: false,
-      hasChildren: false,
-      children: [],
-      expanded: props.defaultopen,
-    };
-  }
-
-  updateWithChildren = (event, data) => {
-    if (event === null) {
-      var children = data["rows"].map((row, index) => {
-        return (<ListChildComponent
-                  id={row["id"]}
-                  name={row["name_translated"]}
-                  changeid={this.props.changeid}
-                  registerinfo={this.props.registerinfo}
-                  getinfo={this.props.getinfo}
-                  expands={true}
-                  defaultopen={false}
-                  key={index}
-                  headnode={false}
-                />);
-      })
-      this.setState({
-        loadedChildren: true,
-        children: children,
-      });
-    } else {
-      console.log("Error: children lookup failed with code " + event.ToString());
-    }
-  }
-
-  loadChildren = () => {
-    // Prevent ourselves from loading children if we've already loaded children
-    if (this.state.loadedChildren || !this.state.hasChildren) {
-      return;
-    }
-
-    // Determine the children of this node
-    var id = this.props.id;
-    var escapedId = id.replace(":", "%5C%3A");
-    var URL = "https://services.phenotips.org/rest/vocabularies/hpo/suggest?sort=nameSort%20asc&maxResults=10000&input=" + id
-            + "&customFilter=is_a:" + escapedId;
-    var xhr = window.Sling.getXHR();
-    xhr.open('GET', URL, true);
-    xhr.responseType = 'json';
-    xhr.onload = () => {
-      var status = xhr.status;
-      if (status === 200) {
-        this.updateWithChildren(null, xhr.response);
-      } else {
-        this.updateWithChildren(status, xhr.response);
-      }
-    };
-    xhr.send();
-  }
-
-  updateChildrenStatus = (event, data) => {
-    if (event === null) {
-      this.setState({
-        hasChildren: (data["rows"].length > 0),
-        checkedForChildren: true,
-      });
-    } else {
-      console.log("Error: children lookup failed with code " + event.ToString());
-    }
-  }
-
-  checkForChildren = () => {
-    // Prevent ourselves for checking for children if we've already checked for children
-    if (this.state.checkedForChildren) {
-      return;
-    }
-
-    // Determine if this node has children
-    var id = this.props.id;
-    var escapedId = id.replace(":", "%5C%3A");
-    var URL = "https://services.phenotips.org/rest/vocabularies/hpo/suggest?sort=nameSort%20asc&maxResults=10000&input=" + id
-            + "&customFilter=is_a:" + escapedId;
-    var xhr = window.Sling.getXHR();
-    xhr.open('GET', URL, true);
-    xhr.responseType = 'json';
-    xhr.onload = () => {
-      var status = xhr.status;
-      if (status === 200) {
-        this.updateChildrenStatus(null, xhr.response);
-      } else {
-        this.updateChildrenStatus(status, xhr.response);
-      }
-    };
-    xhr.send();
-  }
-
-  render() {
-    const { classes, id, name, changeid, registerinfo, getinfo, expands, headnode, bolded } = this.props;
-    if (expands) {
-      this.checkForChildren();
-      if (this.state.expanded) {
-        this.loadChildren();
-      }
-    }
-
-    return(
-      <div key={id} className={headnode ? "" : classes.branch}>
-        {/* Expand button ▼ */}
-        {(expands && this.state.hasChildren) ?
-          <Button
-            onClick={() => {this.setState({expanded: !this.state.expanded})}}
-            variant="text"
-            simple={true}
-            color="info"
-            className={classes.browseitem + " " + classes.arrowDiv}
-            >
-            {this.state.expanded ? "▼" : "▶"}
-          </Button>
-          : ""
-        }
-
-        {/* Listitem button */}
-        <Button
-          onClick={() => changeid(id)}
-          variant="text"
-          simple={true}
-          color="info"
-          className={classes.browseitem}
-          >
-          <Typography inline className={classes.infoDataSource}>{id}&nbsp;</Typography>
-          <Typography inline className={classes.infoName + (bolded ? (" " + classes.boldedName) : " ")}> {name}</Typography>
-        </Button>
-
-        {/* Button to open info page */}
-        <Button
-          buttonRef={(node) => {registerinfo(id, node)}}
-          color="info"
-          justIcon={true}
-          simple={true}
-          onClick={() => {getinfo(id)}}
-          className={classes.buttonLink + " " + classes.infoButton}
-        >
-          <Info color="primary" />
-        </Button>
-        <br />
-
-        {/* Children */}
-        <div className={classes.childDiv + ((expands && this.state.expanded) ? " " : (" " + classes.hiddenDiv)) }> {this.state.children} </div>
-      </div>
-    );
-  }
-}
-
-const ListChildComponent = withStyles(browseStyle)(ListChild);
 
 class BrowseDialog extends React.Component {
   constructor(props) {
@@ -205,7 +46,7 @@ class BrowseDialog extends React.Component {
   // Construct a branch element for rendering
   constructBranch = (id, name, ischildnode, defaultexpanded, bolded) => {
     return(
-      <ListChildComponent
+      <BrowseListChild
         id={id}
         name={name}
         changeid={this.props.changeid}
@@ -358,11 +199,13 @@ class BrowseDialog extends React.Component {
             x
           </Button>
         </div>
-        <div className={classes.treeRoot}>
-          {this.state.parentNode}
-        </div>
-        <div className={classes.treeNode}>
-          {this.state.currentNode}
+        <div className={classes.treeContainer}>
+          <div className={classes.treeRoot}>
+            {this.state.parentNode}
+          </div>
+          <div className={classes.treeNode}>
+            {this.state.currentNode}
+          </div>
         </div>
       </Dialog>
     );
